Extract header to its own file
